Move safe apollo functions to :apollo module
Enables their use in other modules which are not only the :app one.
