+++ apollo/src/main/kotlin/com/hedvig/android/apollo/OperationResult.kt
@@ -1,11 +1,5 @@
-package com.hedvig.app.util.apollo
+package com.hedvig.android.apollo
 
-import arrow.core.Either
-import arrow.core.None
-import arrow.core.Option
-import arrow.core.Some
-import com.apollographql.apollo3.api.ApolloResponse
-import com.apollographql.apollo3.api.Operation
 import com.hedvig.android.core.common.await
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
@@ -81,37 +75,6 @@ sealed interface OperationResult<out T> {
         }
       }
     }
-  }
-}
-
-fun <T> OperationResult<T>.toOption(): Option<T> = when (this) {
-  is OperationResult.Error -> None
-  is OperationResult.Success -> Some(this.data)
-}
-
-fun <T> OperationResult<T>.toEither(): Either<OperationResult.Error, T> = when (this) {
-  is OperationResult.Error -> Either.Left(this)
-  is OperationResult.Success -> Either.Right(this.data)
-}
-
-inline fun <ErrorType, T> OperationResult<T>.toEither(
-  ifEmpty: (message: String?) -> ErrorType,
-): Either<ErrorType, T> = when (this) {
-  is OperationResult.Error -> Either.Left(ifEmpty(message))
-  is OperationResult.Success -> Either.Right(this.data)
-}
-
-fun <D : Operation.Data> ApolloResponse<D>.toOperationResult(): OperationResult<D> {
-  val data = data
-  return when {
-    hasErrors() -> {
-      val exception = errors?.first()?.extensions?.get("exception")
-      val body = (exception as? Map<*, *>)?.get("body")
-      val message = (body as? Map<*, *>)?.get("message") as? String
-      OperationResult.Error.OperationError(message ?: errors?.first()?.message)
-    }
-    data != null -> OperationResult.Success(data)
-    else -> OperationResult.Error.NoDataError("No data")
   }
 }
